codeinstinct/requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.5
distlib==0.3.6
Django==4.2.2
docopt==0.6.2
filelock==3.12.2
identify==2.5.24
idna==3.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
pipreqs==0.4.13
platformdirs==3.8.1
pre-commit==3.3.3
psycopg2==2.9.6
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.0
urllib3==2.0.3
virtualenv==20.23.1
yarg==0.1.9

djangorestframework~=3.14.0